client/utils/ reorganization
Move REST-related types and functions out of client/utils
into a new client/rest package.
