+++ client/rpc/status.go
@@ -5,6 +5,8 @@ import (
 	"net/http"
 	"strconv"
 
+	"github.com/cosmos/cosmos-sdk/client/rest"
+
 	"github.com/spf13/cobra"
 
 	"github.com/spf13/viper"
@@ -12,7 +14,6 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/context"
-	"github.com/cosmos/cosmos-sdk/client/utils"
 )
 
 // StatusCommand returns the status of the network
@@ -77,7 +78,7 @@ func NodeInfoRequestHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 		}
 
 		nodeInfo := status.NodeInfo
-		utils.PostProcessResponse(w, cdc, nodeInfo, cliCtx.Indent)
+		rest.PostProcessResponse(w, cdc, nodeInfo, cliCtx.Indent)
 	}
 }
